feat: use serverless functions via netlify
