+++ client/src/config/index.js
@@ -2,10 +2,12 @@ const port = process.env.PORT || 4000;
 
 const origins = {
     cloud: 'https://face-the-music.eu-gb.cf.appdomain.cloud',
+    netlify: 'https://magicdeck.netlify.app/.netlify/functions/api',
     local: `http://localhost:${port}`,
   };
 
-const serverLocation = process.env.REACT_APP_SERVER_LOCATION || 'local';
+const serverLocation = 'netlify';
+// const serverLocation = process.env.REACT_APP_SERVER_LOCATION || 'local';
 
 const origin = origins[serverLocation];
