+++ package.json
@@ -32,6 +32,7 @@
     "dotenv": "^8.1.0",
     "express": "^4.17.1",
     "geolib": "^3.0.4",
-    "method-override": "^3.0.0"
+    "method-override": "^3.0.0",
+    "serverless-http": "^2.4.1"
   }
 }
